Added window events + Layers + LayersStack

+++ Cajo/src/Cajo/Events/MouseEvent.h
@@ -55,7 +55,7 @@ namespace Cajo {
 	class CAJO_API MouseButtonEvent : public Event
 	{
 	public:
-		inline float GetMouseButton() const { return m_Button; }
+		inline int GetMouseButton() const { return m_Button; }
 
 		EVENT_CLASS_CATEGORY(EventCategoryMouse | EventCategoryInput)
